big change

+++ server/controllers/user_controller.js
@@ -2,14 +2,16 @@ const userModel = require('../models/user_model')
 const activityModel = require("../models/activity_model")
 const messageModel = require('../models/message_model');
 const conversationModel = require('../models/conversation_model');
-const rolesList = require('../configs/roles_config')
+const rolesList = require('../configs/roles_config');
+const { use } = require('../routers/auth_router');
 const readUsers = async (req, res) => {
     try {
 
         const users = await userModel.find({  })
         
         return res.status(200).json({
-            users
+            message:"Users retrieved",
+            res:users
         })
     } catch (error) {
         console.log(error);
@@ -21,18 +23,21 @@ const readUser = async (req, res) => {
 
         if (!id) {
             return res.status(400).json({
-                error: "All inputs are required"
+                message:"All fields are required",
+                res:null
             })
         }
 
         const user = await userModel.findOne({ _id: id })
         if (!user) {
             return res.status(409).json({
-                error: "User does not exists"
+                message:"User does not exists",
+                res:null
             })
         }
         return res.status(200).json({
-            user
+            message:"User retrieved",
+            res:user
         })
     } catch (error) {
         console.log(error);
@@ -45,7 +50,8 @@ const updateUser = async (req, res) => {
         const { name, gender, gender_interest, bio, interests, phone, year} = req.body
         if (!(name && gender && bio && phone && year && gender_interest )) {
             return res.status(400).json({
-                error: "All fields are required"
+                message:"All fields are required",
+                res:null
             })
         }
         const user = await userModel.findByIdAndUpdate({ _id: id }, {
@@ -57,7 +63,10 @@ const updateUser = async (req, res) => {
             gender_interest, 
             interests,
         })
-        return res.status(201).json({user})
+        return res.status(201).json({
+            message:"User updated",
+            res:user
+        })
     } catch (error) {
         console.log(error);
     }
@@ -66,6 +75,12 @@ const updateUser = async (req, res) => {
 const deleteUser = async (req, res) => {
     try {
         const id = req.params.id
+        if (!id) {
+            return res.status(400).json({
+                message:"All fields are required",
+                res:null
+            })
+        }
         const deletedUser = await userModel.deleteOne({ _id: id })
         
         await activityModel.deleteMany({giver_id:id})
@@ -77,7 +92,10 @@ const deleteUser = async (req, res) => {
 
         await conversationModel.deleteMany({members:{$in:[id]}});
 
-        res.status(204).json({ deletedUser })
+        res.status(204).json({
+            message:"User deleted",
+            res:deletedUser
+        })
     } catch (error) {
         console.log(error);
     }
@@ -92,7 +110,10 @@ const deleteUsers = async (req, res) => {
         await conversationModel.deleteMany({});
         await messageModel.deleteMany({});
 
-        res.status(204).json({ deletedUsers })
+        res.status(204).json({
+            message:"All users deleted",
+            res:deletedUsers
+        })
         
     } catch (error) {
         console.log(error);
